[refactor] home component re export
