+++ src/components/Common/index.ts
@@ -1,3 +1,4 @@
+export * from './filter';
 export { default as ContentSubmitButton } from './ContentSubmitButton';
 export { default as GraySubmitButton } from './GraySubmitButton';
 export { default as Layout } from './Layout';
@@ -5,4 +6,3 @@ export { default as Logo } from './Logo';
 export { default as LookTabItem } from './LookTabItem';
 export { default as LookTabs } from './LookTabs';
 export { default as ScrollToTop } from './ScrollToTop';
-export * from './Filter';
