+++ src/components/Home/index.ts
@@ -1,0 +1,3 @@
+export { default as Main } from './Main';
+export * from './Confirm';
+export * from './Today';
