+++ src/page/today/Temperatures.tsx
@@ -1,4 +1,4 @@
-import Main from 'components/Home/Main';
+import { Main } from 'components/Home';
 import { temperatureCategory } from 'constant/lookTab';
 import React from 'react';
